add script to lay out drivers based on their catalog files
This script recursively discovers the driver packages (that is, a
directory containing an .inf file, an .cat file, and possibly other
files) in the source directory.  For each such package, it parses the
catalog file and determines what architectures and windows flavors this
driver is suitable for.  Then it copies or links the driver into
appropriate subdirectories of the destination directory.

As a byproduct it also verifies signatures where it can.

The resulting structure is supposed to be convenient for consupmtion by
both humans and programs.

Note that if there happen to be multiple packages of the same driver
which claim compatibility with a particular architecture and windows
flavor, the one with the most recent timestamp takes over (where the
timestamp is either the signing time, if present, or the catalog
creation time).

Signed-off-by: Roman Kagan <rkagan@virtuozzo.com>