filemap: Update 2k3/x86 build paths
Balloon, NetKVM, and vioserial XP drivers were packaged for both
XP and Windows Server 2003 x86. This saves a little space in the
ISO and there are no significant differences between the two, so
this is more of an OCD fix to make sure that all 2k3 drivers are
TargetOSVersion 5.2 and really "built for Windows 2003".