filemap: Add qemupciserial, qemufwcfg, smbus
qemupciserial was previously manually committed and not handled by
add-whql-submission.py at all. This commit adds support for this
driver as well as the newly added qemufwcfg and smbus.

Note that the drivers consist only of .inf and .cat, there is no .sys.
The drivers come out of the prewhql build as platform / bitness
neutral (qemupciserial under /rhel, qemufwcfg and smbus in the root) and
the same .inf may be signed for different OSes as part of the WHQL
process.

Signed-off-by: Ladi Prosek <lprosek@redhat.com>
Signed-off-by: Cole Robinson <crobinso@redhat.com>