Support new mingw-qemu_ga_win scheme
mingw-qemu_ga_win_changed a while ago to be shipped as RPM instead
of a zip file.

For a while we were regenerating the zip file manually, but it makes
things a little bit more complex (and manual).

This patch allow us to use the rpm package directly, extracting it
to the right folder and including the right msi's in the binary package.

Signed-off-by: Danilo C. L. de Paula <ddepaula@redhat.com>