Add osinfo-db drop-in files
Let's take advantage that osinfo-db supports drop-in files and
ensure osinfo-db consumers will take advantage of the signed
drivers for Windows drivers for Windows 7, Windows 8, Windows 8.1,
and Windows 10.

The current status of this work is rather manual, but right now
seems to be enough for our downstream usage, as agreed with Vadim
in a mail thread.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>